refactor: move AtomixLogStorage classes to broker
