+++ broker/src/main/java/io/camunda/zeebe/broker/system/partitions/PartitionStartupAndTransitionContextImpl.java
@@ -13,6 +13,7 @@ import io.camunda.zeebe.broker.PartitionListener;
 import io.camunda.zeebe.broker.exporter.repo.ExporterDescriptor;
 import io.camunda.zeebe.broker.exporter.repo.ExporterRepository;
 import io.camunda.zeebe.broker.exporter.stream.ExporterDirector;
+import io.camunda.zeebe.broker.logstreams.AtomixLogStorage;
 import io.camunda.zeebe.broker.logstreams.LogDeletionService;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
 import io.camunda.zeebe.broker.system.partitions.impl.AsyncSnapshotDirector;
@@ -24,7 +25,6 @@ import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecordProcessorFa
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.CommandResponseWriter;
 import io.camunda.zeebe.engine.state.QueryService;
 import io.camunda.zeebe.logstreams.log.LogStream;
-import io.camunda.zeebe.logstreams.storage.atomix.AtomixLogStorage;
 import io.camunda.zeebe.snapshots.ConstructableSnapshotStore;
 import io.camunda.zeebe.snapshots.ReceivableSnapshotStore;
 import io.camunda.zeebe.util.health.HealthMonitor;
@@ -371,10 +371,6 @@ public class PartitionStartupAndTransitionContextImpl
     return typedRecordProcessorsFactory::createTypedStreamProcessor;
   }
 
-  public ExporterRepository getExporterRepository() {
-    return exporterRepository;
-  }
-
   @Override
   public ConcurrencyControl getConcurrencyControl() {
     return concurrencyControl;
@@ -383,5 +379,9 @@ public class PartitionStartupAndTransitionContextImpl
   @Override
   public void setConcurrencyControl(final ConcurrencyControl concurrencyControl) {
     this.concurrencyControl = concurrencyControl;
+  }
+
+  public ExporterRepository getExporterRepository() {
+    return exporterRepository;
   }
 }
