+++ broker/src/test/java/io/camunda/zeebe/broker/logstreams/AtomixLogStorageRule.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.logstreams.util;
+package io.camunda.zeebe.broker.logstreams;
 
 import static org.mockito.Mockito.spy;
 
@@ -20,16 +20,12 @@ import io.atomix.raft.zeebe.EntryValidator;
 import io.atomix.raft.zeebe.ValidationResult;
 import io.atomix.raft.zeebe.ZeebeLogAppender;
 import io.camunda.zeebe.logstreams.storage.LogStorage;
-import io.camunda.zeebe.logstreams.storage.atomix.AtomixAppenderSupplier;
-import io.camunda.zeebe.logstreams.storage.atomix.AtomixLogStorage;
-import io.camunda.zeebe.logstreams.storage.atomix.AtomixReaderFactory;
 import io.camunda.zeebe.util.FileUtil;
 import java.io.File;
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.ByteBuffer;
 import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 import java.util.function.LongConsumer;
 import java.util.function.Supplier;
@@ -38,7 +34,7 @@ import org.junit.rules.ExternalResource;
 import org.junit.rules.TemporaryFolder;
 
 public final class AtomixLogStorageRule extends ExternalResource
-    implements AtomixReaderFactory, AtomixAppenderSupplier, ZeebeLogAppender, Supplier<LogStorage> {
+    implements AtomixReaderFactory, ZeebeLogAppender, Supplier<LogStorage> {
 
   private final TemporaryFolder temporaryFolder;
   private final int partitionId;
@@ -53,30 +49,11 @@ public final class AtomixLogStorageRule extends ExternalResource
   private Consumer<Throwable> writeErrorListener;
   private final EntryValidator entryValidator;
 
-  public AtomixLogStorageRule(final TemporaryFolder temporaryFolder) {
-    this(temporaryFolder, 0);
-  }
-
-  public AtomixLogStorageRule(final TemporaryFolder temporaryFolder, final int partitionId) {
-    this(
-        temporaryFolder,
-        partitionId,
-        UnaryOperator.identity(),
-        (a, b) -> ValidationResult.success());
-  }
-
   public AtomixLogStorageRule(
       final TemporaryFolder temporaryFolder,
       final int partitionId,
       final EntryValidator entryValidator) {
     this(temporaryFolder, partitionId, UnaryOperator.identity(), entryValidator);
-  }
-
-  public AtomixLogStorageRule(
-      final TemporaryFolder temporaryFolder,
-      final int partitionId,
-      final UnaryOperator<RaftStorage.Builder> builder) {
-    this(temporaryFolder, partitionId, builder, (a, b) -> ValidationResult.success());
   }
 
   public AtomixLogStorageRule(
@@ -137,27 +114,9 @@ public final class AtomixLogStorageRule extends ExternalResource
     }
   }
 
-  public IndexedRaftLogEntry appendEntry(
-      final long lowestPosition, final long highestPosition, final ByteBuffer data) {
-    final NoopListener listener = new NoopListener();
-    appendEntry(lowestPosition, highestPosition, data, listener);
-
-    return listener.lastWrittenEntry;
-  }
-
   @Override
   public AtomixLogStorage get() {
     return storage;
-  }
-
-  @Override
-  public Optional<ZeebeLogAppender> getAppender() {
-    return Optional.of(this);
-  }
-
-  public CompletableFuture<Void> compact(final long index) {
-    raftLog.deleteUntil(index);
-    return CompletableFuture.completedFuture(null);
   }
 
   @Override
@@ -219,42 +178,7 @@ public final class AtomixLogStorageRule extends ExternalResource
     return partitionId;
   }
 
-  public AtomixLogStorage getStorage() {
-    return storage;
-  }
-
-  public RaftStorage getRaftStorage() {
-    return raftStorage;
-  }
-
-  public RaftLog getRaftLog() {
-    return raftLog;
-  }
-
-  public MetaStore getMetaStore() {
-    return metaStore;
-  }
-
   private RaftStorage.Builder buildDefaultStorage() {
     return RaftStorage.builder().withFlushExplicitly(true).withJournalIndexDensity(1);
-  }
-
-  private static final class NoopListener implements AppendListener {
-
-    private IndexedRaftLogEntry lastWrittenEntry;
-
-    @Override
-    public void onWrite(final IndexedRaftLogEntry indexed) {
-      lastWrittenEntry = indexed;
-    }
-
-    @Override
-    public void onWriteError(final Throwable error) {}
-
-    @Override
-    public void onCommit(final IndexedRaftLogEntry indexed) {}
-
-    @Override
-    public void onCommitError(final IndexedRaftLogEntry indexed, final Throwable error) {}
   }
 }
